Unify hvl_replay.c hvl_tables.c and hvl_tables.h  * hvl2wav had floating point version of the table generators  * Windows port now uses the floating point version of table generators  * hvl2wav had one case label out of position  * hvl2wav had an different variation of white-noise, now in use (see comments)
Main editor still uses its own version of replay.c/replay.h
