+++ Replayer_Amiga/hvl_replay.h
@@ -1,6 +1,7 @@
 #ifndef _REPLAYER_AMIGA_HVL_REPLAY_H
 #define _REPLAYER_AMIGA_HVL_REPLAY_H
 
-#include "../hvl_replay.h"
+#include "types.h"
+#include "../common/hvl_replay.h"
 
 #endif
