+++ Replayer_Amiga/hvl_tables.c
@@ -7,4 +7,4 @@
 #include "hvl_tables.h"
 
 #define HVL_TABLES_FIXEDPOINT
-#include "../hvl_tables.c"
+#include "../common/hvl_tables.c"
