+++ Replayer_Amiga/types.h
@@ -1,0 +1,8 @@
+#ifndef _REPLAYER_AMIGA_TYPES_H
+#define _REPLAYER_AMIGA_TYPES_H
+
+#ifndef const
+#define const
+#endif
+
+#endif
