+++ Replayer_Windows/hvl_replay.h
@@ -1,19 +1,7 @@
 #ifndef _REPLAYER_WINDOWS_HVL_REPLAY_H
 #define _REPLAYER_WINDOWS_HVL_REPLAY_H
 
-typedef char int8;
-typedef unsigned char uint8;
-typedef short int16;
-typedef unsigned short uint16;
-typedef int int32;
-typedef unsigned int uint32;
-typedef double float64;
-typedef char TEXT;
-typedef int BOOL;
-
-#define TRUE 1
-#define FALSE 0
-
-#include "../hvl_replay.h"
+#include "types.h"
+#include "../common/hvl_replay.h"
 
 #endif
