+++ Replayer_Windows/hvl_tables.h
@@ -1,6 +1,7 @@
 #ifndef _REPLAYER_WINDOWS_HVL_TABLES_H
 #define _REPLAYER_WINDOWS_HVL_TABLES_H
 
-#include "../hvl_tables.h"
+#include "types.h"
+#include "../common/hvl_tables.h"
 
 #endif
