+++ Replayer_Windows/types.h
@@ -1,0 +1,22 @@
+#ifndef _REPLAYER_WINDOWS_TYPES_H
+#define _REPLAYER_WINDOWS_TYPES_H
+
+typedef char int8;
+typedef unsigned char uint8;
+typedef short int16;
+typedef unsigned short uint16;
+typedef int int32;
+typedef unsigned int uint32;
+typedef double float64;
+typedef char TEXT;
+typedef int BOOL;
+
+#define TRUE 1
+#define FALSE 0
+
+#ifndef
+#define CONST const
+#endif
+
+
+#endif
