+++ common/hvl_replay.h
@@ -1,6 +1,10 @@
 #ifndef _HVL_REPLAY_H
 #define _HVL_REPLAY_H
 
+/* const is used in general places
+ * CONST is used in places that AMIGA replayer had it
+ */
+
 // Woohoo!
 #define MAX_CHANNELS 16
 
@@ -217,7 +221,7 @@ struct hvl_tune
 void hvl_DecodeFrame( struct hvl_tune *ht, int8 *buf1, int8 *buf2, int32 bufmod );
 void hvl_InitReplayer( void );
 BOOL hvl_InitSubsong( struct hvl_tune *ht, uint32 nr );
-struct hvl_tune *hvl_LoadTune( const TEXT *name, uint32 freq, uint32 defstereo );
+struct hvl_tune *hvl_LoadTune( CONST TEXT *name, uint32 freq, uint32 defstereo );
 struct hvl_tune *hvl_ParseTune(const uint8 *buf, uint32 buflen, uint32 freq, uint32 defstereo ); /* not all hvl_replay.c implementations provide this yet */
 void hvl_FreeTune( struct hvl_tune *ht );
