+++ hvl2wav/hvl_tables.h
@@ -2,6 +2,6 @@
 #define _HVL2WAV_HVL_TABLES_H
 
 #include "types.h"
-#include "../hvl_tables.h"
+#include "../common/hvl_tables.h"
 
 #endif
